+++ forms-n-views/Code/Java/frostillicus/dxl/Folder.java
@@ -1,8 +1,15 @@
 package frostillicus.dxl;
 
+import java.io.*;
 import java.util.*;
 import javax.xml.xpath.XPathExpressionException;
+
+import lotus.domino.*;
+
+import com.ibm.xsp.extlib.util.ExtLibUtil;
 import com.raidomatic.xml.*;
+
+import frostillicus.FNVUtil;
 
 public class Folder extends AbstractDXLDesignNote {
 	private static final long serialVersionUID = -8774232556021141733L;
@@ -45,5 +52,36 @@ public class Folder extends AbstractDXLDesignNote {
 	public void swapColumns(int a, int b) throws XPathExpressionException {
 		XMLNodeList columnNodes = (XMLNodeList)getDxl().selectNodes("//column");
 		columnNodes.swap(a, b);
+	}
+
+	public static String create(String databaseDocumentId, String name) throws Exception {
+		DxlImporter importer = null;
+		try {
+			InputStream is = Stylesheet.class.getResourceAsStream("/dxl/folder.xml");
+			BufferedReader reader = new BufferedReader(new InputStreamReader(is));
+			StringBuilder xmlBuilder = new StringBuilder();
+			while(reader.ready()) {
+				xmlBuilder.append(reader.readLine());
+				xmlBuilder.append("\n");
+			}
+			is.close();
+			String xml = xmlBuilder.toString().replace("name=\"\"", "name=\"" + FNVUtil.xmlEncode(name) + "\"");
+
+			importer = ExtLibUtil.getCurrentSession().createDxlImporter();
+			importer.setDesignImportOption(DxlImporter.DXLIMPORTOPTION_REPLACE_ELSE_CREATE);
+			importer.setReplicaRequiredForReplaceOrUpdate(false);
+			Document databaseDoc = ExtLibUtil.getCurrentDatabase().getDocumentByUNID(databaseDocumentId);
+			Database foreignDB = ExtLibUtil.getCurrentSessionAsSignerWithFullAccess().getDatabase(databaseDoc.getItemValueString("Server"), databaseDoc.getItemValueString("FilePath"));
+			importer.importDxl(xml, foreignDB);
+
+			Document importedDoc = foreignDB.getDocumentByID(importer.getFirstImportedNoteID());
+			return importedDoc.getUniversalID();
+		} catch(Exception e) {
+			e.printStackTrace();
+			if(importer != null) {
+				System.out.println(importer.getLog());
+			}
+		}
+		return null;
 	}
 }